employer/recruiter_app_fixed_v2/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
sqlalchemy==2.0.5
python-dotenv==1.0.0
pydantic==1.10.7
